Update file logic
