+++ src/PageContainer/MainPage/style.ts
@@ -1,0 +1,46 @@
+/** @jsxImportSource @emotion/react */
+import styled from '@emotion/styled';
+
+export const ArrowBox = styled.div({
+  position: 'absolute',
+  top: '60.75rem',
+  width: '3.75rem',
+  height: '1.375rem',
+  left: '98.125rem',
+  display: 'flex',
+  justifyContent: 'space-between',
+  alignItems: 'center',
+});
+
+export const Arrow = styled.div({
+  display: 'flex',
+  justifyContent: 'space-between',
+  alignItems: 'center',
+});
+
+export const Comment = styled.p({
+  position: 'absolute',
+  width: '21.875rem',
+  height: '2.0625rem',
+  left: '18.3125rem',
+  top: '58.125rem',
+  fontSize: '1.75rem',
+});
+
+export const DotsBox = styled.div({
+  display: 'flex',
+  width: '6.25rem',
+  height: '0.625rem',
+  position: 'absolute',
+  top: '87.375rem',
+  left: '56.8125rem',
+});
+
+export const Dots = styled.div({
+  display: 'flex',
+  width: '0.625rem',
+  height: '0.625rem',
+  backgroundColor: '#D9D9D9',
+  margin: '0rem 0.1875rem',
+  borderRadius: '100%',
+});
