+++ src/PageContainer/StartPage/style.ts
@@ -1,0 +1,31 @@
+/** @jsxImportSource @emotion/react */
+import styled from '@emotion/styled';
+
+export const Container = styled.div({
+  position: 'absolute',
+  width: '100%',
+  height: '800px',
+  left: '0px',
+  top: '120px',
+  background: '#DDE8F8',
+  display: 'flex',
+  justifyContent: 'center',
+  alignItems: 'center',
+});
+
+export const Description = styled.div({
+  position: 'absolute',
+  width: '285px',
+  height: '57px',
+  left: '260px',
+  top: '332px',
+  color: '#383838',
+});
+
+export const Title = styled.div({
+  position: 'absolute',
+  width: '577px',
+  height: '86px',
+  left: '260px',
+  top: '224px',
+});
