+++ src/PageContainer/index.tsx
@@ -1,0 +1,2 @@
+export { default as MainPage } from './MainPage';
+export { default as StartPage } from './StartPage';
