+++ src/components/KakaoMap/index.tsx
@@ -15,27 +15,22 @@ const KakaoMap = ({ text, svg }: { text: string; svg: any }) => {
     {
       title: '카카오',
       latlng: { lat: 33.450705, lng: 126.570677 },
-      id: 1,
     },
     {
       title: '생태연못',
       latlng: { lat: 33.450936, lng: 126.569477 },
-      id: 2,
     },
     {
       title: '텃밭',
       latlng: { lat: 33.450879, lng: 126.56994 },
-      id: 3,
     },
     {
       title: '근린공원',
       latlng: { lat: 33.451393, lng: 126.570738 },
-      id: 4,
     },
     {
       title: '광주소프트웨어마이스터고',
       latlng: { lat: 35.142738601752846, lng: 126.80072297715732 },
-      id: 5,
     },
   ];
   return (
@@ -48,8 +43,8 @@ const KakaoMap = ({ text, svg }: { text: string; svg: any }) => {
             lng: 126.80072297715732,
           }}
           style={{
-            width: '1352px',
-            height: '480px',
+            width: '84.5rem',
+            height: '30rem',
           }}
           level={1}
         >
