+++ src/components/KakaoMap/style.ts
@@ -1,0 +1,67 @@
+/** @jsxImportSource @emotion/react */
+import styled from '@emotion/styled';
+
+export const MapComment = styled.p({
+  position: 'absolute',
+  width: '26.875rem',
+  height: '2.0625rem',
+  left: '18.3125rem',
+  top: '8.125rem',
+  fontSize: '1.75rem',
+});
+
+export const Map = styled.div({
+  width: '84.5rem',
+  height: '30rem',
+  position: 'absolute',
+  top: '16.8125rem',
+  left: '17.75rem',
+  backgroundColor: 'black',
+});
+
+export const Wrap = styled.div({
+  position: 'absolute',
+  display: 'flex',
+  justifyContent: 'center',
+  alignItems: 'center',
+  width: '18.75rem',
+  height: '6rem',
+  borderRadius: '0.5rem',
+  backgroundColor: 'white',
+  bottom: '90px',
+  right: '-150px',
+});
+
+export const Overlay = styled.div({
+  display: 'flex',
+  justifyContent: 'space-evenly',
+  alignItems: 'center',
+  width: '18.75rem',
+});
+export const JibunEllipsis = styled.div({
+  fontSize: '12px',
+  color: '#999999',
+});
+export const Ellipsis = styled.div({
+  fontSize: '12px',
+  color: '#999999',
+});
+export const DescTitle = styled.div({
+  fontSize: '14px',
+  paddingBottom: '8px',
+});
+export const Desc = styled.div({
+  display: 'flex',
+  flexDirection: 'column',
+  justifyContent: 'space-around',
+  height: '49px',
+  paddingBottom: '5px',
+});
+export const LocationImg = styled.img({
+  borderRadius: '8px',
+});
+export const Esc = styled.div({
+  position: 'absolute',
+  left: '280px',
+  top: '5px',
+});
