--- src/components/MainBox/style.ts
@@ -1,91 +1,0 @@
-/** @jsxImportSource @emotion/react */
-import styled from '@emotion/styled';
-
-export const Container = styled.div({
-  position: 'absolute',
-  width: '1920px',
-  height: '800px',
-  left: '0px',
-  top: '120px',
-  background: '#DDE8F8',
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-});
-
-export const StartBtn = styled.div({
-  width: '80px',
-  height: '44px',
-  left: '260px',
-  top: '549px',
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-});
-
-export const Description = styled.div({
-  position: 'absolute',
-  width: '285px',
-  height: '57px',
-  left: '260px',
-  top: '332px',
-  color: '#383838',
-  // font-family: 'Pretendard',
-  // font-style: 'normal',
-  // font-weight: '500',
-  // font-size: '22px',
-  // line-height: '26px',
-  // letter-spacing: '-0.01em',
-});
-
-export const Title = styled.div({
-  position: 'absolute',
-  width: '577px',
-  height: '86px',
-  left: '260px',
-  top: '224px',
-});
-
-export const StartBtnBox = styled.div({
-  // box-sizing: 'border-box',
-  position: 'absolute',
-  width: '120px',
-  height: '44px',
-  left: '260px',
-  top: '429px',
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-  border: ' 1px solid #000000',
-  borderRadius: '8px',
-  boxSizing: 'border-box',
-  // border: '1px solid #000000',
-  // border-radius: '8px',
-});
-
-export const MarkerBox = styled.div({
-  position: 'absolute',
-  transform: 'rotate(-6deg)',
-  height: '461px',
-  width: '363px',
-  left: '1016.9px',
-  top: '126.29px',
-  borderRadius: '0px',
-});
-
-export const MarkerBox2 = styled.div({
-  position: 'absolute',
-  transform: 'rotate(-6deg)',
-  height: '461px',
-  width: '363px',
-  left: '1412px',
-  top: '284px',
-  borderRadius: '0px',
-});
-
-export const Shadow = styled.div({
-  position: 'absolute',
-  right: '20px',
-});
-
-export const Shadow2 = styled.div({});
