+++ src/components/Marker1/style.ts
@@ -1,0 +1,17 @@
+/** @jsxImportSource @emotion/react */
+import styled from '@emotion/styled';
+
+export const MarkerBox = styled.div({
+  position: 'absolute',
+  transform: 'rotate(-6deg)',
+  height: '461px',
+  width: '363px',
+  left: '1016.9px',
+  top: '126.29px',
+  borderRadius: '0px',
+});
+
+export const Shadow = styled.div({
+  position: 'absolute',
+  right: '20px',
+});
