+++ src/components/Marker2/index.tsx
@@ -1,7 +1,7 @@
 /** @jsxImportSource @emotion/react */
 import { css } from '@emotion/react';
 import * as S from './style';
-export const MarkerBox2 = ({ svg, svg2 }: { svg: any; svg2: any }) => {
+const Marker2 = ({ svg, svg2 }: { svg: any; svg2: any }) => {
   return (
     <S.MarkerBox2>
       {svg}
@@ -9,3 +9,5 @@ export const MarkerBox2 = ({ svg, svg2 }: { svg: any; svg2: any }) => {
     </S.MarkerBox2>
   );
 };
+
+export default Marker2;
