+++ src/components/Marker2/style.ts
@@ -1,0 +1,14 @@
+/** @jsxImportSource @emotion/react */
+import styled from '@emotion/styled';
+
+export const MarkerBox2 = styled.div({
+  position: 'absolute',
+  transform: 'rotate(-6deg)',
+  height: '461px',
+  width: '363px',
+  left: '1412px',
+  top: '284px',
+  borderRadius: '0px',
+});
+
+export const Shadow2 = styled.div({});
