+++ src/components/PlusBtn/style.ts
@@ -1,0 +1,15 @@
+/** @jsxImportSource @emotion/react */
+import styled from '@emotion/styled';
+
+export const PlusBtn = styled.button({
+  width: '5rem',
+  height: '5rem',
+  backgroundColor: '#2E6EEB',
+  position: 'absolute',
+  top: '90.5rem',
+  left: '107.25rem',
+  borderRadius: '100%',
+  border: 'none',
+});
+
+export const Plus = styled.div({});
