+++ src/components/StartBtn/index.tsx
@@ -1,7 +1,7 @@
 /** @jsxImportSource @emotion/react */
 import { css } from '@emotion/react';
 import * as S from './style';
-export const StartBtnBox = ({ svg, text }: { svg: any; text: string }) => {
+const StartBtn = ({ svg, text }: { svg: any; text: string }) => {
   return (
     <S.StartBtnBox>
       <S.StartBtn>{text}</S.StartBtn>
@@ -9,3 +9,5 @@ export const StartBtnBox = ({ svg, text }: { svg: any; text: string }) => {
     </S.StartBtnBox>
   );
 };
+
+export default StartBtn;
