+++ src/components/StartBtn/style.ts
@@ -1,0 +1,29 @@
+/** @jsxImportSource @emotion/react */
+import styled from '@emotion/styled';
+
+export const StartBtn = styled.div({
+  width: '80px',
+  height: '44px',
+  left: '260px',
+  top: '549px',
+  display: 'flex',
+  justifyContent: 'center',
+  alignItems: 'center',
+});
+
+export const StartBtnBox = styled.div({
+  // box-sizing: 'border-box',
+  position: 'absolute',
+  width: '120px',
+  height: '44px',
+  left: '260px',
+  top: '429px',
+  display: 'flex',
+  justifyContent: 'center',
+  alignItems: 'center',
+  border: ' 1px solid #000000',
+  borderRadius: '8px',
+  boxSizing: 'border-box',
+  // border: '1px solid #000000',
+  // border-radius: '8px',
+});
