+++ src/components/index.ts
@@ -1,2 +1,6 @@
-export { default as MainBox } from './MainBox';
+export { default as PlusBtn } from './PlusBtn';
+export { default as KakaoMap } from './KakaoMap';
 export { default as Preview } from './Preview';
+export { default as StartBtn } from './StartBtn';
+export { default as Marker1 } from './Marker1';
+export { default as Marker2 } from './Marker2';
