+++ src/pages/MainPage.tsx
@@ -1,6 +1,11 @@
-import { Preview } from '../components';
-const MainPage = () => {
-  return <Preview></Preview>;
+import { MainPage } from 'PageContainer';
+import * as I from 'assets/svgs';
+const Main = () => {
+  return (
+    <>
+      <MainPage svg={<I.ArrowIcon />}></MainPage>
+    </>
+  );
 };
 
-export default MainPage;
+export default Main;
